excel-ai-interviewer/requirements.txt
streamlit
fastapi
sqlalchemy
google-generativeai
python-dotenv
pandas
openpyxl